test(cli): split snapshot mutation target coverage
